ml-deploy-platform/backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
scikit-learn
joblib
numpy
pandas
google-genai
python-dotenv
